Move crate_engine stuff into its own file
